Assignment 1
Reflection

In this first assignment, i can learn how to make a simple programming. The importance of programming is that it helps us solve problems and carry out tasks more efficiently. It also allows us to automate processes and create new ways of doing things. Computer programming is a highly sought-after skill in an increasingly technological world.